requirements.txt
playwright==1.46.0
Pillow==10.4.0
tldextract==5.1.2
